get Code, Cost, Data
These are my 3 get functions, i made a variable at the start with the
filename so it is easy to change later on, and for my getData function
it wanted me to return a string and a long, so i made a string vector so
it will return the description and the cost both as strings.  If you
have comments let me know!